move rerandomized test to frost-rerandomized; clean up dependencies; add 'frost' feature
