+++ Cargo.toml
@@ -30,7 +30,7 @@ pasta_curves = { version = "0.4", default-features = false }
 rand_core = { version = "0.6", default-features = false }
 serde = { version = "1", optional = true, features = ["derive"] }
 thiserror = { version = "1.0", optional = true }
-frost-rerandomized = { git = "https://github.com/ZcashFoundation/frost.git", rev = "4e20d20135ae1b149a64678eff3651b9cc8bb574", optional = true }
+frost-rerandomized = { git = "https://github.com/ZcashFoundation/frost.git", rev = "7cabe0feb108aa5f52fcead45d323266c45efe10", optional = true }
 
 [dependencies.zeroize]
 version = "1"
@@ -46,9 +46,7 @@ proptest = "1.0"
 rand = "0.8"
 rand_chacha = "0.3"
 serde_json = "1.0"
-frost-rerandomized = { git = "https://github.com/ZcashFoundation/frost.git", rev = "4e20d20135ae1b149a64678eff3651b9cc8bb574" }
-frost-core = { git = "https://github.com/ZcashFoundation/frost.git", rev = "4e20d20135ae1b149a64678eff3651b9cc8bb574", features=["test-impl"]  }
-
+frost-rerandomized = { git = "https://github.com/ZcashFoundation/frost.git", rev = "7cabe0feb108aa5f52fcead45d323266c45efe10", features=["test-impl"] }
 
 # `alloc` is only used in test code
 [dev-dependencies.pasta_curves]
@@ -59,6 +57,7 @@ std = ["blake2b_simd/std", "thiserror", "zeroize", "alloc",
        "serde", "frost-rerandomized"] # conditional compilation for serde not complete (issue #9)
 alloc = []
 nightly = []
+frost = ["frost-rerandomized"]
 default = ["std"]
 
 [[bench]]
