+++ tests/frost_redpallas.rs
@@ -1,25 +1,46 @@
-use rand::thread_rng;
-use reddsa::frost_redpallas::PallasBlake2b512;
+#![cfg(feature = "frost")]
 
-mod frost;
+use rand::thread_rng;
+
+use reddsa::{frost_redpallas::PallasBlake2b512, orchard};
 
 #[test]
 fn check_sign_with_dealer() {
     let rng = thread_rng();
 
-    frost_core::tests::check_sign_with_dealer::<PallasBlake2b512, _>(rng);
+    frost_rerandomized::frost_core::tests::check_sign_with_dealer::<PallasBlake2b512, _>(rng);
 }
 
 #[test]
 fn check_randomized_sign_with_dealer() {
     let rng = thread_rng();
 
-    frost::check_randomized_sign_with_dealer::<PallasBlake2b512, _>(rng);
+    let (msg, group_signature, group_pubkey) =
+        frost_rerandomized::tests::check_randomized_sign_with_dealer::<PallasBlake2b512, _>(rng);
+
+    // Check that the threshold signature can be verified by the `reddsa` crate
+    // public key (interoperability test)
+
+    let sig = {
+        let bytes: [u8; 64] = group_signature.to_bytes().as_ref().try_into().unwrap();
+        reddsa::Signature::<orchard::SpendAuth>::from(bytes)
+    };
+    let pk_bytes = {
+        let bytes: [u8; 32] = group_pubkey.to_bytes().as_ref().try_into().unwrap();
+        reddsa::VerificationKeyBytes::<orchard::SpendAuth>::from(bytes)
+    };
+
+    // Check that the verification key is a valid RedDSA verification key.
+    let pub_key = reddsa::VerificationKey::try_from(pk_bytes)
+        .expect("The test verification key to be well-formed.");
+
+    // Check that signature validation has the expected result.
+    assert!(pub_key.verify(&msg, &sig).is_ok());
 }
 
 #[test]
 fn check_sign_with_dkg() {
     let rng = thread_rng();
 
-    frost_core::tests::check_sign_with_dkg::<PallasBlake2b512, _>(rng);
+    frost_rerandomized::frost_core::tests::check_sign_with_dkg::<PallasBlake2b512, _>(rng);
 }
